+++ Assets/Scripts/Enemy/Controllers/EnemyHealthController.cs
@@ -4,7 +4,7 @@ namespace Assets.Scripts.Enemy.Controllers
 {
     public class EnemyHealthController : UnitHealthController
     {
-        public override void HealthModify(float value)
+        public override void HealthModify(int value)
         {
             unitModel.ModifyHealth(value);
         }
